Early Withdrawal / Early Repayment

+++ contracts/external/WETH9.sol
@@ -27,8 +27,8 @@ contract WETH9 is IERC20 {
     mapping(address => uint256) public override balanceOf;
     mapping(address => mapping(address => uint256)) public override allowance;
 
-    event  Deposit(address indexed dst, uint wad);
-    event  Withdrawal(address indexed src, uint wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
     receive() external payable {
         deposit();
